don't destroy the node after it's processed. add beg/end to note the time a node could be launch, and the time it should be waited for until it's done.
Circular dependency was detected by checking on three conditions:
in_progress=RESET, in_progress=SET, node destroyed
now these are mapped to:
in_progress=RST, in_progress=PRE, in_progress=SET

the in_progress is marked as REQ when a node is required to be done,
in the mean time, variable end is calculated.

Currently, the filename is printed when it's required (as in previous paragraph),
and the result is identical to the original.